relocate: Utility to transform files and functions to URLs
I'd like to add references to files and functions in this article, with
links to the go sources. This includes links to the appropriate line
number when referencing functions.

To do so, let's have a small language in the AsciiDoc file and a
preprocessor that resolve them to the correct github URL. Of course, for
functions, it relies on parsing the file and find out the line where the
function is defined. Fortunately, go provides all that is needed.

The tool currently recognises:

  @@path/to/file@@		// relative to the root of the go repo
  @@path/to/file:function@@

The preprocessor will error out is a file or function doesn't exist
anymore, prompting an update to the doc.

Signed-off-by: Damien Lespiau <damien.lespiau@intel.com>